Teams are created based on bases. More than 1 player may be on a team at a time based on how it is configured.
When a player connects whichever team has the fewest number of players
(and is not at their limit) has the Player added to the team. The player
is successfully spawned in the Teams spawn point. Need to adjust the
color of the player (based on the material) when the team is assigned.